Move helper functions into ctapipe.reco submodule
