--- docs/api-reference/vectorization/index.rst
@@ -1,20 +1,0 @@
-.. _vectorization:
-
-*******************************************************************************
-Helper Functions for Vectorized Operations on Tables (`~ctapipe.vectorization`)
-*******************************************************************************
-
-.. currentmodule:: ctapipe.vectorization
-
-This module contains helper functions for performing vectorized operations
-on tables, as used in e.g. :ref:`reco_stereo_combination`.
-
-
-Reference/API
-=============
-
-.. automodapi:: ctapipe.vectorization
-    :no-inheritance-diagram:
-
-.. automodapi:: ctapipe.vectorization.aggregate
-    :no-inheritance-diagram:
