--- src/ctapipe/vectorization/__init__.py
@@ -1,6 +1,0 @@
-from .aggregate import get_subarray_index, weighted_mean_std_ufunc
-
-__all__ = [
-    "get_subarray_index",
-    "weighted_mean_std_ufunc",
-]
